LibGfx/JPEGXL: Move the entropy decoder to its own file
No functional change.
